chore: reorganize project folder structure
