+++ src/store/queries/comment/useCommentMutation.ts
@@ -1,5 +1,5 @@
 import { patchComment } from "@/apis/comment";
-import { TCommentBlockReq } from "@/types/comment";
+import { TCommentBlockReq } from "@/types/comment.type";
 import { TStatusResponse } from "@/types/index.type";
 import { displayErrorDialog } from "@/utils/sweetAlert";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
