--- src/store/zustand/store.ts
@@ -1,1 +1,0 @@
-//store
